feat: make icon folder and remove index key
